Add files via upload
It requires the same work as byungjae89, the only difference is that you need to download the mvtec dataset and unzip it, put it in the mvtec_anomaly_detection folder under the data folder, and then run main.py. I provided the data results I got after training in the test_log.txt file. you can use as reference